AioSession 升级状态机设计

+++ Z-Queen/src/main/java/com/tgx/z/queen/io/core/inf/ISession.java
@@ -53,7 +53,7 @@ public interface ISession
 
     void setIndex(long _Index);
 
-    int getHashKey();
+    long getHashKey();
 
     AsynchronousSocketChannel getChannel();
